pyproject.toml
[tool.poetry]
name = "sencha"
version = "0.1.0"
description = ""
authors = ["Daniel Mauricio Flores <dnielfs@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
mistralai = "^0.2.0"
flask = "^3.0.3"
flask-cors = "^4.0.1"
python-dotenv = "^1.0.1"
openai = "^1.30.3"
elevenlabs = "^1.2.2"
groq = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
